changed filters to trait and added after, before and raw filter for searching messages. thanks @gmarciano #51
